+++ apps/client/src/common/hooks-query/useAppVersion.ts
@@ -2,7 +2,8 @@ import { useQuery } from '@tanstack/react-query';
 import { dayInMs } from 'ontime-utils';
 
 import { version } from '../../../../../package.json';
-import { APP_VERSION, isLocalhost } from '../api/constants';
+import { isLocalhost } from '../../externals';
+import { APP_VERSION } from '../api/constants';
 import { getLatestVersion, HasUpdate } from '../api/external';
 
 const placeholder: HasUpdate & { hasUpdates: boolean } = { url: '', version: '', hasUpdates: false };
