+++ apps/client/src/features/app-settings/panel/shutdown-panel/ShutdownPanel.tsx
@@ -10,8 +10,8 @@ import {
   useDisclosure,
 } from '@chakra-ui/react';
 
-import { isLocalhost } from '../../../../common/api/constants';
 import { useElectronEvent } from '../../../../common/hooks/useElectronEvent';
+import { isLocalhost } from '../../../../externals';
 import * as Panel from '../../panel-utils/PanelUtils';
 
 export default function ShutdownPanel() {
